+++ src/auth.js
@@ -17,6 +17,21 @@ const uuid = require('uuid'),
     logger = require('./logger'),
     url = require('url');
 
+/**
+ *
+ * @param {Object} request         The request Object. Can optionally contain a
+ *                                 'headersToSign' property: An ordered list header names
+ *                                 that will be included in the signature. This will be
+ *                                 provided by specific APIs.
+ * @param {String} clientToken     The client token value from the .edgerc file.
+ * @param {String} accessToken     The access token value from the .edgerc file.
+ * @param {String} clientSecret    The client secret value from the .edgerc file.
+ * @param {Date} timestamp         The timestamp with format "yyyyMMddTHH:mm:ss+0000".
+ * @param {String} nonce           A random string used to detect replayed request messages.
+ * @param {Number} maxBody         This parameter is deprecated.
+ * @returns {string}
+ * @deprecated maxBody
+ */
 function makeAuthHeader(request, clientToken, accessToken, clientSecret, timestamp, nonce, maxBody) {
     const keyValuePairs = {
         client_token: clientToken,
@@ -58,6 +73,22 @@ function makeURL(host, path, queryStringObj) {
 }
 
 module.exports = {
+    /**
+     *
+     * @param {Object} request        The request Object. Can optionally contain a
+     *                                'headersToSign' property: An ordered list header names
+     *                                that will be included in the signature. This will be
+     *                                provided by specific APIs.
+     * @param {String} clientToken    The client token value from the .edgerc file.
+     * @param {String} clientSecret   The client secret value from the .edgerc file.
+     * @param {String} accessToken    The access token value from the .edgerc file.
+     * @param {String} host           The host a unique string followed by luna.akamaiapis.net from the .edgerc file.
+     * @param {Number} maxBody        This value is deprecated.
+     * @param {String} guid           A random string used to detect replayed request messages.
+     * @param {Date} timestamp        The timestamp with format "yyyyMMddTHH:mm:ss+0000".
+     * @returns {{headers}|*}         The request Object.
+     * @deprecated maxBody
+     */
     generateAuth: function (request, clientToken, clientSecret, accessToken, host, maxBody, guid, timestamp) {
         guid = guid || uuid.v4();
         timestamp = timestamp || helpers.createTimestamp();